backend/requirements.txt
fastapi
uvicorn
python-multipart
aiofiles
google-cloud-storage
google-cloud-speech
google-cloud-translate
google-cloud-vision
google-cloud-aiplatform
supabase
python-dotenv
